✨ Refactor imports: move DirectorApi and RabbitMQ references to clients module; add new RabbitMQ client
