+++ services/catalog/src/simcore_service_catalog/api/rest/_services_extras.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends
 from models_library.api_schemas_directorv2.services import ServiceExtras
 from models_library.services import ServiceKey, ServiceVersion
 
-from ...service.director import DirectorApi
+from ...clients.director import DirectorApi
 from .._dependencies.director import get_director_api
 
 router = APIRouter()
