+++ services/catalog/src/simcore_service_catalog/clients/rabbitmq.py
